+++ barapost/barapost.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "3.14.d"
+__version__ = "3.14.e"
 # Year, month, day
-__last_update_date__ = "2020-05-26"
+__last_update_date__ = "2020-06-05"
 
 # |===== Check python interpreter version =====|